Allow custom metadata in data files.
The Iceberg spec requires custom metadata fields for Avro files. It
expects information such as the table scheme, the partition information
and the manifest type to be in the Manifest File's metadata.

This change makes the `setMetadata` method of `DataFileWriterBase`
public instead of private, and adds a similar public method to
`DataFileWriter`.