Add `map` logical type for arrays.
Maps can be represented by either the built-in map type, or an array of
key-value pairs with the logical type `map`. The Iceberg "column_sizes"
field is an example of the latter.

This change allows arrays to be annotated with the `map` logical type,
which was previously not supported in C++ Avro.