allow for iceberg element-ids take 2
Iceberg manifests need to be able to write their JSONified Avro schemas
in the Avro headers for certain Iceberg clients (e.g. PyIceberg) to be
able to read them.

Commit 966d457e725b5decdafdd453934af85999d7ac57 was a stab at allowing
this by hacking this custom field into the AST, but it did so as a
string, instead of a long. This isn't exactly what other libraries
output, but is sufficient for PyIceberg. To avoid surprises down the
line, this quick hack switches us to serialize the long.

(cherry picked from commit b2b82a455001eef51b69e19d608dfa50667e7b49)