C++: add elementId to programmatic API
A previous commit added a "element-id" attribute to arrays, which will
be used in Iceberg manifests. This was previously only available when
parsing a schema JSON file, not via the programmatic API.

The way that Redpanda will build Iceberg manifests is going to change to
use the programmatic API, in order to support some fields of the
manifest that have a dynamic schema. To that end, this exposes the
"element-id" via a new constructor for ArraySchemas.