C++: add default value when building schemas
The default value of a schema is important to know when trying to parse
data records that have missing values.

Avro's JSON-compiling schema builder constructs a GenericDatum from type
the provided in the schema[1]: When there is no default provided in the
schema, the default datum is GenericDatum(). Otherwise, it's a datum of
the type desired by the schema, constructed with values pulled from the
schema.

In my immediate use case (Iceberg manifest entries) most fields have no
default values, but union types will have a default of null, in which
case the caller will need to pass in a GenericDatum containing a
GenericUnion.

This introduces some interfaces to the RecordSchema::addField() method
to allow the addition of defaults to fields.

[1] https://github.com/redpanda-data/avro/blob/3bbe74853f8cd3184272aa7c130b3cab45828f62/lang/c%2B%2B/impl/Compiler.cc#L301